requirements.txt
# Dev environment
pip
autopep8

# App
flask

Flask-SQLAlchemy
Flask-Migrate
python-dotenv
trendspy # pytrends is no longer work
openai